backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
pandas
Pillow
pytesseract
openpyxl
pymupdf
